requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1

chromadb==0.4.24
langchain-chroma==0.1.2

sentence-transformers==2.7.0
torch==2.2.2

pypdf==4.2.0
streamlit==1.33.0
python-dotenv==1.0.1
numpy==1.26.4

google-generativeai>=0.8.0

transformers==4.41.2
accelerate==0.31.0
